+++ packages/hono-takibi/src/core/schema/references/index.ts
@@ -1,6 +1,5 @@
 export { extractRefs } from './extract-refs.js'
 export { findReferences } from './find-references.js'
-export { getRefSchemaName } from './get-ref-schema-name.js'
 export { resolveSchemasDependencies } from './resolve-schemas-dependencies.js'
 export { traverseSchema } from './traverse-schema.js'
 export { traverseSchemaDependencies } from './traverse-schema-dependencies.js'
