+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/components/anyof/index.ts
@@ -1,4 +1,3 @@
-import { getRefSchemaName } from '../../../../../core/schema/references/index.js'
 import type { Schema } from '../../../../../openapi/index.js'
 import { union } from '../../../../zod/z/index.js'
 import { zodToOpenAPI } from '../../../../zod-to-openapi/index.js'
@@ -22,7 +21,7 @@ export function anyOf(schema: Schema): string {
   }
 
   const zodSchemas = schema.anyOf.map((subSchema) => {
-    subSchema.$ref ? getRefSchemaName(subSchema.$ref) : zodToOpenAPI(subSchema)
+    subSchema.$ref ? `${subSchema.$ref}Schema` : zodToOpenAPI(subSchema)
     return zodToOpenAPI(subSchema)
   })
