+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { createRoute, route, routeCode, routeName } from '.'
+import { createRoute, route, routeCode } from '.'
 
 // Test run
 // pnpm vitest run ./src/generator/zod-openapi-hono/openapi/route/index.test.ts
@@ -11,10 +11,6 @@ describe('route module barrel file exports', () => {
 
   it('should export route', () => {
     expect(typeof route).toBe('function')
-  })
-
-  it('should export routeName', () => {
-    expect(typeof routeName).toBe('function')
   })
 
   it('should export routeCode', () => {
