+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/index.ts
@@ -1,4 +1,3 @@
 export { createRoute } from './create-route.js'
 export { route } from './route.js'
 export { routeCode } from './route-code.js'
-export { routeName } from './route-name.js'
