+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/route.ts
@@ -1,7 +1,6 @@
-import { escapeStringLiteral } from '../../../../core/utils/index.js'
+import { escapeStringLiteral, routeName } from '../../../../core/utils/index.js'
 import type { Operation } from '../../../../openapi/index.js'
 import { createRoute } from './create-route.js'
-import { routeName } from './index.js'
 import { requestParameter } from './params/index.js'
 import { response } from './response/index.js'
